Refactor layout handling and update footer styles
- Removed the layout file from the pages directory.
- Created a new Layout component to encapsulate page layouts.
- Introduced media queries and responsive design adjustments.
