+++ src/Screens/HomeScreen/HomeScreen.tsx
@@ -1,9 +1,9 @@
 "use client";
 import {
-	BgImage,
-	IconBottom,
-	MainContainer,
-	Wrapper,
+  BgImage,
+  IconBottom,
+  MainContainer,
+  Wrapper,
 } from "./HomeScreenStyles";
 import { Text } from "@/components/Text/Text";
 import { colors } from "../../../theme/colors";
@@ -23,78 +23,79 @@ import { useAuthStore } from "@/store/AuthStore";
 import { ITEMS } from "./mock";
 
 const HomeScreen = () => {
-	const productStore = useProductStore();
-	const currencyStore = useCurrencyStore();
-	const handleScrollToBottom = () => {
-		if (typeof window !== "undefined") {
-			window.scrollTo({ top: window.innerHeight, behavior: "smooth" });
-		}
-	};
+  const productStore = useProductStore();
+  const currencyStore = useCurrencyStore();
+  const handleScrollToBottom = () => {
+    if (typeof window !== "undefined") {
+      window.scrollTo({ top: window.innerHeight, behavior: "smooth" });
+    }
+  };
 
-	useEffect(() => {
-		productStore.fetchSales();
-		productStore.fetchForYou();
-	}, []);
-	const { t } = useTranslation();
-	const authStore = useAuthStore();
-	const refresh = () => {
-		authStore.refreshToken();
-	};
-	return (
-		<>
-			<UseMetaData title={"Home"} img={""} description={"asd"} />
-			<Wrapper>
-				<BgImage bgImage='/images/home/bannerImage.png'>
-					<MainContainer>
-						<Text
-							color={colors.white}
-							size='54px'
-							margin='20% 0'
-							fontStyle={fonts.f500}
-							preline>
-							{t("home.mainText")}
-						</Text>
-					</MainContainer>
-					<IconBottom
-						alt='Arrow Scroll Bottom'
-						width={50}
-						height={50}
-						src='/icons/mainArrowBottom.svg'
-						onClick={() => handleScrollToBottom()}
-					/>
-				</BgImage>
-				<MainContainer>
-					<ByPropose items={ITEMS} />
+  useEffect(() => {
+    productStore.fetchSales();
+    productStore.fetchForYou();
+  }, []);
+  const { t } = useTranslation();
+  const authStore = useAuthStore();
+  const refresh = () => {
+    authStore.refreshToken();
+  };
+  return (
+    <>
+      <UseMetaData title={"Home"} img={""} description={"asd"} />
+      <Wrapper>
+        <BgImage bgImage="/images/home/bannerImage.png">
+          <MainContainer>
+            <Text
+              color={colors.white}
+              size="54px"
+              margin="20% 0"
+              fontStyle={fonts.f500}
+              preline
+            >
+              {t("home.mainText")}
+            </Text>
+          </MainContainer>
+          <IconBottom
+            alt="Arrow Scroll Bottom"
+            width={50}
+            height={50}
+            src="/icons/mainArrowBottom.svg"
+            onClick={() => handleScrollToBottom()}
+          />
+        </BgImage>
+        <MainContainer>
+          <ByPropose items={ITEMS} />
 
-					<ResponsiveBlockGroup variant='1' />
-					{productStore?.forYou.length && (
-						<Slider
-							title={t("home.forYou")}
-							items={productStore?.forYou}
-							variant='cards'
-						/>
-					)}
-					{productStore?.sales.length && (
-						<Slider
-							title={t("home.popularAcc")}
-							items={productStore?.sales}
-							variant='cards'
-						/>
-					)}
+          <ResponsiveBlockGroup variant="1" />
+          {productStore?.forYou.length && (
+            <Slider
+              title={t("home.forYou")}
+              items={productStore?.forYou}
+              variant="cards"
+            />
+          )}
+          {productStore?.sales.length && (
+            <Slider
+              title={t("home.popularAcc")}
+              items={productStore?.sales}
+              variant="cards"
+            />
+          )}
 
-					<ResponsiveBlockGroup variant='2' />
-					{/*<Slider
+          <ResponsiveBlockGroup variant="2" />
+          {/*<Slider
 						title={"Велоблог"}
 						tags={[]}
 						items={newsData}
 						variant='news'
 					/> */}
-					<ResponsiveBlockGroup variant='3' />
-				</MainContainer>
+          <ResponsiveBlockGroup variant="3" />
+        </MainContainer>
 
-				<Subscribe bg={"/images/home/form-background.png"} />
-			</Wrapper>
-		</>
-	);
+        <Subscribe bg={"/images/home/form-background.png"} />
+      </Wrapper>
+    </>
+  );
 };
 export default observer(HomeScreen);
