+++ src/mock/data.ts
@@ -1,354 +1,363 @@
 import { Product } from "@/types/types";
 
 export const byProposeData = [
-    {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-      {
-        title: 'Title',
-        picture: '/mock/testCardByPropose.png',
-        link: 'asdadasd',
-        count: 2
-      },
-]
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+  {
+    title: "Title",
+    picture: "/mock/testCardByPropose.png",
+    link: "asdadasd",
+    count: 2,
+  },
+];
 
 export const sliderData: Product[] = [
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray', 'red'],
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray", "red"],
+    sale: "30",
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray'],
-    sizes: ['12','14', '16'],
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray"],
+    sizes: ["12", "14", "16"],
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    sale: "30",
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray'],
-    sizes: ['12','14', '16'],
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray"],
+    sizes: ["12", "14", "16"],
+    sale: "30",
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray'],
-    sizes: ['12','14', '16'],
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray"],
+    sizes: ["12", "14", "16"],
+    sale: "30",
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray'],
-    sizes: ['12','14', '16'],
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray"],
+    sizes: ["12", "14", "16"],
+    sale: "30",
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray'],
-    sizes: ['12','14', '16'],
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray"],
+    sizes: ["12", "14", "16"],
+    sale: "30",
   },
   {
-    title: 'Turbo Creo SL Expert',
-    image: '/mock/testCardByPropose.png',
-    price: '47800',
-    colors: ['black', 'gray'],
-    sizes: ['12','14', '16'],
-    sale: '30'
+    title: "Turbo Creo SL Expert",
+    image: "/mock/testCardByPropose.png",
+    price: "47800",
+    colors: ["black", "gray"],
+    sizes: ["12", "14", "16"],
+    sale: "30",
   },
-]
+];
 export const sliderTags = [
-    "Тег 1",
-    "Тег тег 2",
-    "тег3",
-    "те тег4",
-    "ТЕГ 5",
-    "6 тег",
-    "7 Тег 7",
-    "Теггггггг 8",
-    "тег 9 ТЕГ",
-  ];
+  "Тег 1",
+  "Тег тег 2",
+  "тег3",
+  "те тег4",
+  "ТЕГ 5",
+  "6 тег",
+  "7 Тег 7",
+  "Теггггггг 8",
+  "тег 9 ТЕГ",
+];
 
-  export const footerItems = {
-    ua:[{title: 'Гірські ',
-    items: ['Кросс-кантрі',
-       ' Трейл',
-        'Швидкісний спуск',
-        'Дьорт\Памп-трек',
-        'Рами\Фреймсети'
-       ]
-    },
-    {title: 'Шосе ',
-    items: ['Performance',
-       'Гравійні',
-        'Циклокрос',
-        'Тріатлон',
-        'Рами\Фреймсети'
-       ]
-    },
-    {title: 'Active ',
-    items: ['Фитнес',
-       ' Компьютери',
-        'Комфорт',
-       ]
-    },
-    {title: 'ELECTRIC ',
-    items: ['Road',
-       ' E-MTB',
-        'Active',
-       ]
-    },
-    {title: 'Дитячі ',
-    items: ['Від 1.5 до 4 років',
-       'Від 4 до 7 років',
-        'Від 7 до 10 років',
-       ]
-    },
-    {title: 'Знаряддя ',
-    items: ['Взуття',
-       ' Шоломи',
-        'Світло',
-        'Насоси',
-        'Сумки & Зберігання'
-       ]
-    },
-    {title: 'Кліентам ',
-    items: ['Про магазин',
-       'Майстерня',
-        'Велоблог',
-        'Контакти',
-       ]
-    },
-    
-    {title:"Зв'язатись",
-    items: ['(044) 440-22-00',
-       'bikelove@bike.com',
-        
-       ]
-    },],
-    ru:[{title: 'Горные ',
-    items: ['Кросс-кантри',
-       ' Трейл',
-        'Скоростной спуск',
-        'Дёрт\Памп-трек',
-        'Рамы\Фреймсеты'
-       ]
-    },
-    {title: 'Шоссе ',
-    items: ['Performance',
-       'Гравийные',
-        'Циклокросс',
-        'Триатлон',
-        'Рамы\Фреймсеты'
-       ]
-    },
-    {title: 'Active ',
-    items: ['Фитнес',
-       ' Коммьютеры',
-        'Комфорт',
-       ]
-    },
-    {title: 'ELECTRIC ',
-    items: ['Road',
-       ' E-MTB',
-        'Active',
-       ]
-    },
-    {title: 'Детские ',
-    items: ['От 1.5 до 4 лет',
-       'От 4 до 7 лет',
-        'От 7 до 10 лет',
-       ]
-    },
-    {title: 'Экипировка ',
-    items: ['    Обувь',
-       ' Шлемы',
-        'Свет',
-        'Насосы',
-        'Сумки & Хранение'
-       ]
-    },
-    {title: 'Клиентам ',
-    items: ['О магазине',
-       'Мастерская',
-        'Велоблог',
-        'Контакты',
-       ]
-    },
-    
-    {title: 'Связаться ',
-    items: ['(044) 440-22-00',
-       'bikelove@bike.com',
-        
-       ]
-    },],
-    
-  }
-  export const newsData = [
-    {
-      "id": 1,
-      "title": "Оптимизация посадки",
-      "description": "Весь процесс оптимизации посадки сопровождается отслеживанием на мониторе показателей мощности и каденса в режиме реального времени.",
-      "image": "/mock/testNews.png",
-      "date": "20.12.2023"
-    },
-    {
-      "id": 2,
-      "title": "Мониторинг параметров",
-      "description": "Мониторинг показателей мощности и каденса в режиме реального времени способствует оптимизации посадки.",
-      "image": "/mock/testNews.png",
-      "date": "24.12.2023"
-    },
-    {
-      "id": 3,
-      "title": "Отслеживание данных",
-      "description": "Отслеживание данных о мощности и каденсе помогает в процессе оптимизации посадки.",
-      "image": "/mock/testNews.png",
-      "date": "15.12.2023"
-    },
-    {
-      "id": 4,
-      "title": "Оптимизация посадки",
-      "description": "Эффективная оптимизация посадки требует постоянного отслеживания показателей мощности и каденса.",
-      "image": "/mock/testNews.png",
-      "date": "20.12.2023"
-    },
-    {
-      "id": 5,
-      "title": "Улучшение результатов",
-      "description": "Улучшение результатов требует систематического мониторинга мощности и каденса в режиме реального времени.",
-      "image": "/mock/testNews.png",
-      "date": "7.12.2023"
-    },
-    {
-      "id": 6,
-      "title": "Оптимизация посадки",
-      "description": "Отслеживание параметров мощности и каденса на мониторе помогает в оптимизации посадки.",
-      "image": "/mock/testNews.png",
-      "date": "20.12.2023"
-    }
-  ]
-  
+// export const footerItems = {
+//   ua: [
+//     {
+//       title: "Гірські ",
+//       items: [
+//         "Кросс-кантрі",
+//         " Трейл",
+//         "Швидкісний спуск",
+//         "ДьортПамп-трек",
+//         "РамиФреймсети",
+//       ],
+//     },
+//     {
+//       title: "Шосе ",
+//       items: [
+//         "Performance",
+//         "Гравійні",
+//         "Циклокрос",
+//         "Тріатлон",
+//         "РамиФреймсети",
+//       ],
+//     },
+//     { title: "Active ", items: ["Фитнес", " Компьютери", "Комфорт"] },
+//     { title: "ELECTRIC ", items: ["Road", " E-MTB", "Active"] },
+//     {
+//       title: "Дитячі ",
+//       items: ["Від 1.5 до 4 років", "Від 4 до 7 років", "Від 7 до 10 років"],
+//     },
+//     {
+//       title: "Знаряддя ",
+//       items: ["Взуття", " Шоломи", "Світло", "Насоси", "Сумки & Зберігання"],
+//     },
+//     {
+//       title: "Кліентам ",
+//       items: ["Про магазин", "Майстерня", "Велоблог", "Контакти"],
+//     },
 
-  export const CATEGORIES_TITLES = [
-    "Велосипеды",
-    "Электрический транспорт",
-    "Компоненты",
-    "Аксессуары",
-    "Одежда",
-    "Экипировка",
-    "Инструменты и обслуживание",
-  ];
-  export const CATEGORIES = [
+//     { title: "Зв'язатись", items: ["(044) 440-22-00", "bikelove@bike.com"] },
+//   ],
+//   ru: [
+//     {
+//       title: "Горные ",
+//       items: [
+//         "Кросс-кантри",
+//         " Трейл",
+//         "Скоростной спуск",
+//         "ДёртПамп-трек",
+//         "РамыФреймсеты",
+//       ],
+//     },
+//     {
+//       title: "Шоссе ",
+//       items: [
+//         "Performance",
+//         "Гравийные",
+//         "Циклокросс",
+//         "Триатлон",
+//         "РамыФреймсеты",
+//       ],
+//     },
+//     { title: "Active ", items: ["Фитнес", " Коммьютеры", "Комфорт"] },
+//     { title: "ELECTRIC ", items: ["Road", " E-MTB", "Active"] },
+//     {
+//       title: "Детские ",
+//       items: ["От 1.5 до 4 лет", "От 4 до 7 лет", "От 7 до 10 лет"],
+//     },
+//     {
+//       title: "Экипировка ",
+//       items: ["    Обувь", " Шлемы", "Свет", "Насосы", "Сумки & Хранение"],
+//     },
+//     {
+//       title: "Клиентам ",
+//       items: ["О магазине", "Мастерская", "Велоблог", "Контакты"],
+//     },
+
+//     { title: "Связаться ", items: ["(044) 440-22-00", "bikelove@bike.com"] },
+//   ],
+// };
+export const newsData = [
+  {
+    id: 1,
+    title: "Оптимизация посадки",
+    description:
+      "Весь процесс оптимизации посадки сопровождается отслеживанием на мониторе показателей мощности и каденса в режиме реального времени.",
+    image: "/mock/testNews.png",
+    date: "20.12.2023",
+  },
+  {
+    id: 2,
+    title: "Мониторинг параметров",
+    description:
+      "Мониторинг показателей мощности и каденса в режиме реального времени способствует оптимизации посадки.",
+    image: "/mock/testNews.png",
+    date: "24.12.2023",
+  },
+  {
+    id: 3,
+    title: "Отслеживание данных",
+    description:
+      "Отслеживание данных о мощности и каденсе помогает в процессе оптимизации посадки.",
+    image: "/mock/testNews.png",
+    date: "15.12.2023",
+  },
+  {
+    id: 4,
+    title: "Оптимизация посадки",
+    description:
+      "Эффективная оптимизация посадки требует постоянного отслеживания показателей мощности и каденса.",
+    image: "/mock/testNews.png",
+    date: "20.12.2023",
+  },
+  {
+    id: 5,
+    title: "Улучшение результатов",
+    description:
+      "Улучшение результатов требует систематического мониторинга мощности и каденса в режиме реального времени.",
+    image: "/mock/testNews.png",
+    date: "7.12.2023",
+  },
+  {
+    id: 6,
+    title: "Оптимизация посадки",
+    description:
+      "Отслеживание параметров мощности и каденса на мониторе помогает в оптимизации посадки.",
+    image: "/mock/testNews.png",
+    date: "20.12.2023",
+  },
+];
+
+export const CATEGORIES_TITLES = [
+  "Велосипеды",
+  "Электрический транспорт",
+  "Компоненты",
+  "Аксессуары",
+  "Одежда",
+  "Экипировка",
+  "Инструменты и обслуживание",
+];
+export const CATEGORIES = [
+  {
+    title: "Категория",
+    categories: [
+      {
+        title: "Подкатегория",
+        categories: [
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+        ],
+      },
+      {
+        title: "Подкатегория",
+        categories: [
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+        ],
+      },
+      {
+        title: "Подкатегория",
+        categories: [
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+        ],
+      },
+      {
+        title: "Подкатегория",
+        categories: [
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+          "Подкатегория",
+        ],
+      },
+    ],
+  },
+];
+
+export const textData = {};
+
+export const sliderGallery = [
+  { image: "/images/workshop/gallery.png", text: "Регул юей скор" },
+  { image: "/images/workshop/gallery.png", text: "Регул юей скор" },
+  { image: "/images/workshop/gallery.png", text: "Регул юей скор" },
+  { image: "/images/workshop/gallery.png", text: "Регул юей скор" },
+  { image: "/images/workshop/gallery.png", text: "Регул юей скор" },
+  { image: "/images/workshop/gallery.png", text: "Регул юей скор" },
+];
+
+export const footerItems = {
+  ua: [
     {
-      title: "Категория",
-      categories: [
-        {
-          title: "Подкатегория",
-          categories: [
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-          ],
-        },
-        {
-          title: "Подкатегория",
-          categories: [
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-          ],
-        },
-        {
-          title: "Подкатегория",
-          categories: [
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-          ],
-        },
-        {
-          title: "Подкатегория",
-          categories: [
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-            "Подкатегория",
-          ],
-        },
+      title: "Інформація",
+      items: ["Про нас", "Доставка та оплата", "Контакти", "Публічна оферта"],
+    },
+    {
+      title: "Режим роботи",
+      items: [
+        "ПН 10:00 - 20:00",
+        "ВТ 10:00 - 20:00",
+        "СР 10:00 - 20:00",
+        "ЧТ 10:00 - 20:00",
+        "ПТ 10:00 - 20:00",
+        "СБ 10:00 - 20:00",
+        "НД 10:00 - 20:00",
       ],
     },
-  ];
-
-  export const textData = {
-    
-  }
-
-  export const sliderGallery = [
-    {image: '/images/workshop/gallery.png', text: 'Регул юей скор'},
-    {image: '/images/workshop/gallery.png', text: 'Регул юей скор'},
-    {image: '/images/workshop/gallery.png', text: 'Регул юей скор'},
-    {image: '/images/workshop/gallery.png', text: 'Регул юей скор'},
-    {image: '/images/workshop/gallery.png', text: 'Регул юей скор'},
-    {image: '/images/workshop/gallery.png', text: 'Регул юей скор'},
-
-  ]
+  ],
+  ru: [
+    {
+      title: "Информация",
+      items: [
+        "О нас",
+        "Доставка и оплата",
+        "Контакты",
+        "Публичное предложение",
+      ],
+    },
+    {
+      title: "Режим работы",
+      items: [
+        "ПН 10:00 - 20:00",
+        "ВТ 10:00 - 20:00",
+        "СР 10:00 - 20:00",
+        "ЧТ 10:00 - 20:00",
+        "ПТ 10:00 - 20:00",
+        "СБ 10:00 - 20:00",
+        "ВС 10:00 - 20:00",
+      ],
+    },
+  ],
+};
