+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import Footer from "@/components/Footer/Footer";
+import Footer from "@/components/Layout/Footer/Footer";
 import "@/styles/globals.css";
 import { Auth0Provider } from "@auth0/auth0-react";
 import type { AppProps } from "next/app";
@@ -6,40 +6,42 @@ import NextNProgress from "nextjs-progressbar";
 import { CONFIG } from "../../config";
 import { UseMetaData } from "@/helpers/hooks/useMetaData";
 import { ToastContainer } from "react-toastify";
-import Layout from "./layout";
+
 import PhoneWidget from "@/components/PhoneWidget/PhoneWidget";
 import { I18nextProvider } from "react-i18next";
 import i18n from "@/i18n/i18n";
+import Layout from "@/components/Layout";
 
 export default function App({ Component, pageProps }: AppProps) {
-	return (
-		<>
-			<ToastContainer
-				position='top-center'
-				autoClose={3500}
-				hideProgressBar={true}
-				newestOnTop={true}
-				closeOnClick
-			/>
-			<Auth0Provider
-				domain={CONFIG.AUTH0.domain}
-				clientId={CONFIG.AUTH0.clientId}
-				authorizationParams={{ redirect_uri: CONFIG.AUTH0.redirect }}
-				cacheLocation='localstorage'>
-				<Layout>
-					<NextNProgress
-						color='red'
-						startPosition={0}
-						stopDelayMs={200}
-						height={3}
-						showOnShallow={true}
-					/>
-					<I18nextProvider i18n={i18n}>
-						<Component {...pageProps} />
-					</I18nextProvider>
-					<PhoneWidget />
-				</Layout>
-			</Auth0Provider>
-		</>
-	);
+  return (
+    <>
+      <ToastContainer
+        position="top-center"
+        autoClose={3500}
+        hideProgressBar={true}
+        newestOnTop={true}
+        closeOnClick
+      />
+      <Auth0Provider
+        domain={CONFIG.AUTH0.domain}
+        clientId={CONFIG.AUTH0.clientId}
+        authorizationParams={{ redirect_uri: CONFIG.AUTH0.redirect }}
+        cacheLocation="localstorage"
+      >
+        <Layout>
+          <NextNProgress
+            color="red"
+            startPosition={0}
+            stopDelayMs={200}
+            height={3}
+            showOnShallow={true}
+          />
+          <I18nextProvider i18n={i18n}>
+            <Component {...pageProps} />
+          </I18nextProvider>
+          <PhoneWidget />
+        </Layout>
+      </Auth0Provider>
+    </>
+  );
 }
